+++ src/config/project/types.ts
@@ -10,4 +10,7 @@ export interface ProjectConstants {
   maxContractNameLength: number;
   maxContractDescriptionLength: number;
   maxCodeNameLength: number;
+
+  // move verify task
+  maxMoveVerifyTaskRequestNoteLength: number;
 }