+++ src/lib/app-provider/hooks/useFabricateFee.ts
@@ -1,16 +1,13 @@
 import type { StdFee } from "@cosmjs/stargate";
 import { useCallback } from "react";
 
-import { useCelatoneApp } from "../contexts/app";
 import { big } from "lib/types";
 import type { Gas } from "lib/types";
 
+import { useGas } from "./useGas";
+
 export const useFabricateFee = () => {
-  const {
-    chainConfig: {
-      gas: { gasPrice, gasAdjustment, maxGasLimit },
-    },
-  } = useCelatoneApp();
+  const { tokenPerGas, denom, gasAdjustment, maxGasLimit } = useGas();
 
   return useCallback(
     (estimatedGas: number): StdFee => {
@@ -22,13 +19,13 @@ export const useFabricateFee = () => {
       return {
         amount: [
           {
-            denom: gasPrice.denom,
-            amount: big(adjustedGas).mul(gasPrice.tokenPerGas).toFixed(0),
+            denom,
+            amount: big(adjustedGas).mul(tokenPerGas).toFixed(0),
           },
         ],
         gas: adjustedGas.toString() as Gas<string>,
       };
     },
-    [gasAdjustment, gasPrice.denom, gasPrice.tokenPerGas, maxGasLimit]
+    [gasAdjustment, denom, tokenPerGas, maxGasLimit]
   );
 };