+++ src/lib/components/icon/UploadIcon.tsx
@@ -1,5 +1,6 @@
 import { Flex } from "@chakra-ui/react";
 
+import { CustomIcon } from "./CustomIcon";
 import { SvgIcon } from "./SvgIcon";
 
 interface UploadIconProps {
@@ -28,6 +29,22 @@ export const UploadIcon = ({ variant = "primary" }: UploadIconProps) => {
       bgColor={style.bgColor}
     >
       <SvgIcon name="upload" boxSize="24px" color={style.color} />
+    </Flex>
+  );
+};
+
+export const UploadFolderIcon = ({ variant = "primary" }: UploadIconProps) => {
+  const style = getVariantStyle(variant);
+  return (
+    <Flex
+      align="center"
+      justify="center"
+      borderRadius="50%"
+      w="40px"
+      h="40px"
+      bgColor={style.bgColor}
+    >
+      <CustomIcon name="folder" boxSize={4} color={style.color} />
     </Flex>
   );
 };