+++ src/lib/model/code.ts
@@ -2,28 +2,25 @@ import { useMemo } from "react";
 
 import { useCurrentChain } from "lib/app-provider";
 import type { PermissionFilterValue } from "lib/hooks";
-import {
-  useCodePermissionFilter,
-  useCodeSearchFilter,
-  useUserKey,
-} from "lib/hooks";
+import { useCodePermissionFilter, useCodeSearchFilter } from "lib/hooks";
 import { useCodeStore } from "lib/providers/store";
 import {
+  useAllCodesByAddress,
   useCodeListByCodeIds,
-  useCodeListByWalletAddress,
 } from "lib/services/wasm/code";
-import type { CodeInfo } from "lib/types";
+import type { BechAddr, CodeInfo } from "lib/types";
 import { AccessConfigPermission } from "lib/types";
 
 const useStoredCodes = () => {
   const { address } = useCurrentChain();
   const { getCodeLocalInfo, isCodeIdSaved, isHydrated } = useCodeStore();
 
-  const { data: rawStoredCodes, isLoading } =
-    useCodeListByWalletAddress(address);
+  const { data: rawStoredCodes, isLoading } = useAllCodesByAddress(
+    address as BechAddr
+  );
 
   const storedCodes =
-    rawStoredCodes?.map<CodeInfo>((code) => ({
+    rawStoredCodes?.items.map<CodeInfo>((code) => ({
       ...code,
       name: getCodeLocalInfo(code.id)?.name,
       isSaved: isCodeIdSaved(code.id),
@@ -33,13 +30,12 @@ const useStoredCodes = () => {
 };
 
 const useSavedCodes = () => {
-  const userKey = useUserKey();
   const { lastSavedCodes, lastSavedCodeIds, isHydrated } = useCodeStore();
 
-  const savedCodeIds = lastSavedCodeIds(userKey);
+  const savedCodeIds = lastSavedCodeIds();
   const { data: rawSavedCodes, isLoading } = useCodeListByCodeIds(savedCodeIds);
 
-  const savedCodes = lastSavedCodes(userKey).map<CodeInfo>((localSavedCode) => {
+  const savedCodes = lastSavedCodes().map<CodeInfo>((localSavedCode) => {
     const rawSavedCode = rawSavedCodes?.find(
       (savedCode) => savedCode.id === localSavedCode.id
     );