+++ src/lib/pages/proposal-details/components/VoteQuorumBadge.tsx
@@ -1,11 +1,11 @@
 import { isNull } from "lodash";
 
 import { PeriodState } from "../types";
+import { ActiveDot } from "lib/components/ActiveDot";
 import { CustomIcon } from "lib/components/icon";
 import type { Nullable, Ratio } from "lib/types";
 import { ProposalStatus } from "lib/types";
 
-import { ActiveDot } from "./ActiveDot";
 import { ProgressBadge } from "./ProgressBadge";
 
 interface VoteQuorumBadgeProps {